부산대 BE_박정우 6주차 과제(2단계) (#90)
* fix : 필터 경로 수정

* feat : cors 해결 코드 구현

* feat : 에러 코드 협의후 정리, 인증 로직 정리

* fix : 예외처리 에러 수정

* feat : cors 전체허용

* fix : cors 설정변경

* fix : cors 수정

* fix : cors 문제 해결

* fix : cors 수정

* fix : corsfilter 추가

* fix : cors filter 다시 삭제

* fix : 리다이랙션 에러 반환 로직으로 변경

* feat : 프리플라이트 헤더 설정 추가

* fix : 에러 반환하도록 변경

* fix : filter 변경

* fix : filter 변경

* fix : jwt로 주문하기로 수정

* fix : 리뷰사항 반영

* fix : filter 해제

+++ src/main/java/gift/service/WishService.java
@@ -9,6 +9,7 @@ import gift.domain.Wish;
 import gift.repository.ProductRepository;
 import gift.repository.UserInfoRepository;
 import gift.repository.WishRepository;
+import gift.utils.error.AuthorizationException;
 import gift.utils.error.ProductNotFoundException;
 import gift.utils.error.UserNotFoundException;
 import gift.utils.error.WishListNotFoundException;
@@ -50,9 +51,17 @@ public class WishService {
     @Transactional
     public WishResponse removeFromWishlist(String email, Long wishId) {
 
+        UserInfo userInfo = userInfoRepository.findByEmail(email).orElseThrow(
+            () -> new UserNotFoundException("User Not Found")
+        );
+
         Wish wish = wishRepository.findById(wishId).orElseThrow(
             () -> new WishListNotFoundException("Wish Not Found")
         );
+
+        if (userInfo.getId().equals(wish.getProduct().getId())){
+            throw new AuthorizationException("Not Authorization");
+        }
 
         wishRepository.deleteByProductIdAndUserInfoId(wish.getProduct().getId(), wish.getUserInfo().getId());
         return new WishResponse(wish.getId(),wish.getProduct().getId(),wish.getCount());
